lambda/StravaStatsChanger/requirements.txt
stravainteractweblib==0.0.5
PyJWT==2.6.0